pyproject.toml
[tool.poetry]
name = "tango-mlflow"
version = "0.1.2"
description = ""
authors = ["altescy <altescy@fastmail.com>"]

[tool.poetry.scripts]
tango-mlflow = "tango_mlflow.__main__:run"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
ai2-tango = "^1.0.0"
mlflow = ">=1.28,<3.0"
optuna = {version = "^3.0.5", optional = true}
llvmlite = "^0.39.1"
numba = "^0.56.4"
torch = {version = "^2.0.0", optional = true}
flax = {version = "^0.6.8", optional = true}

[tool.poetry.dev-dependencies]
python-language-server = "^0.36.2"
pytest = "^7.1.3"
pysen = {version = "^0.10.2"}
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
mypy = "^0.991"

[tool.poetry.extras]
optuna = ["optuna"]
torch = ["torch"]
flax = ["flax"]
all = ["optuna", "torch", "flax"]

[tool.poetry.group.dev.dependencies]
jax = {version = "^0.4.8", extras = ["cpu"]}
datasets = "^2.11.0"

[tool.pysen]
version = "0.10"

[tool.pysen-cli]
settings_dir = "."

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 120
py_version = "py38"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.pysen.lint.source]
  includes = ["."]
  excludes = [".venv/"]


[tool.black] # automatically generated by pysen
# pysen ignores and overwrites any modifications
line-length = 120
target-version = ["py38"]

[tool.isort] # automatically generated by pysen
# pysen ignores and overwrites any modifications
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 120
multi_line_output = 3
use_parentheses = true
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
